Fix the leave group button in Internet Explorer
The way the leave group button works in every other browser is that it
uses the `form` attribute (and also `formmethod`) so that, even though
the <button> isn't inside the search bar <form> in the HTML, it belongs
to that form. That means that any search text that the user has typed
into the search <input> is also submitted when you click the leave group
button, because the input and button are part of the same form.

Note that although un-lozengified search terms are preserved, any search
terms in the search bar <input> that have been "lozengified" by
JavaScript code will, however, be lost when the leave group button is
clicked. This is because the JavaScript lozengification actually removes
the search terms from the <input> element and so removes them from the
form. This has to be fixed in JavaScript, in a follow-up commit.

Unfortunately no version of IE supports the `form` attribute so for IE
it's necessary for the leave group button to have its own `<form>`.
Because this is a `<form method="POST">` any search query terms that are
already in the URL will be preserved when the button is clicked.
Unfortunately, unlike in other browsers, there is no way (without
JavaScript) to preserve any text that has been typed into the search
bar's <input> but not submitted to the server (and so not part of the
URL) yet, because the search <input> and the leave group <button> belong
to different forms. Even un-lozengified text in the <input> is lost in
IE.

I think this is the best that can be done for IE without JavaScript, but
we can add some JavaScript in a follow-up commit so that unsubmitted
search terms will not be lost even in IE.