Add "More info" / "Back" nav to group search page
Add a navigation component showing the group name and a "More info"
button to the top of the groups search page on small screens.

This nav only shows on smaller screens, after the group info sidebar has
moved below the search results.

The nav shows a "More info" button that, when clicked, loads a version
of the page in which the nav shows a "Back" button instead.

This is, of course, pointless. But in the future the only search results
(and no sidebar) will be shown on the page with the "More info" button
and only the sidebar will be shown in the page with the "Back" button,
so the nav will toggle between two different pages.

Known issues: any search terms are lost when the button is clicked.