Update hidden input live
Change SearchBarController to insert its hidden input into the DOM at
controller initialization and keep its value up to date "live" as text
is lozengified, lozenges are deleted, and text is entered into the
visible <input>.

Previously SearchBarController didn't generate the value for the hidden
input, and didn't insert the hidden input into the form, until the form
was submitted.

The problem with this is that we now want to add various new submit
buttons to this form that are outside of SearchBarController's element,
and those submit buttons would end up submitting an empty q value as
SearchBarController would have removed search text from the <input> and
turned it into lozenges, but since SearchBarController was not
submitting the form it would not insert its hidden input (containing the
values from the lozenges).

As long as SearchBarController always keeps the value of its hidden
input up to date then the form can be submitted in the normal way
(without JavaScript) at any time by any submit button and the full value
of the query will be submitted.

SearchBarController no longer does anything at submit time.